\name{refB}
\docType{data}
\alias{refB}
\title{ MAQC reference data }
\description{ A subset of the MAQC's 'B' RNA reference dataset }
\details{
The Microarray Quality Control Consortium (MAQC) has generated a reference 
dataset of Human Genome U133 Plus 2.0 GeneChips from 100\% of Ambion Brain 
Reference RNA. Five different test (\_1\_ to \_6\_) sites have processed 
this reference in 5 replicates (B1 to B5). 

The refB dataset includes one randomly chosen array per test site against
which the user's hgu133plus2.0 array will be compared. See the package 
vignette for more details.
}
\format{an object of type \code{"AffyBatch"}}
\seealso{refA, refC, refD}
\keyword{datasets}
